更新

+++ application/admin/model/AdminUser.php
@@ -90,10 +90,10 @@ class AdminUser extends Model
             $this->error = '请指定需要删除的用户ID！';
             return false;
         }
-        /*if ($userId == config('USER_ADMINISTRATOR')) {
-        $this->error = '禁止对超级管理员执行该操作！';
-        return false;
-        }*/
+        if ($userId == 1) {
+            $this->error = '禁止对超级管理员执行该操作！';
+            return false;
+        }
         if (false !== $this->where(array('userid' => $userId))->delete()) {
             return true;
         } else {
